requirements.txt
pytorch_lightning
torch
torchvision
numpy
einops
tensorboard
tensorboardX
plotly
pandas
scikit-learn
tqdm
fire
mido
pillow
tqdm